categories connected to eachredux
