+++ src/Lib/api-calls.js
@@ -1,5 +1,7 @@
-export let getAllIdeas = () =>
-    fetch("https://mysterious-everglades-33844.herokuapp.com/", {
+export let getAllIdeas = () => {
+    return (
+        fetch("https://mysterious-everglades-33844.herokuapp.com/", {
         method: "GET",
         mode: "cors"
     })
+)}
